textpb: Add ToValue methods to Message and Value.
These methods render the message into Go map and slice values.  Maps are from
string to interface{} and the concrete type of each value reflects its
structure, i.e., another map for a submessage, a primitive value for a string
or number, or a slice of interface{} values for an array.